requirements.txt
pyyaml
psycopg2
elasticsearch
flake8
